// src/types/order.type.ts
// {
//   items: [
//     { meal_id: '7e09890c-d1fd-4de5-9bdc-d13336cfbfc8', quantity: 2 },
//     { meal_id: '425244fa-d630-47ef-bb90-5f68fb3c596c', quantity: 2 }
//   ]
// }

export type OrderPayload = {
  items: {
    meal_id: string;
    provider_id: string;
    quantity: number;
  }[],
  delivery_address: string;
};

export type MealItem = {
  meal_id: string;
  provider_id: string;
  quantity: number;
};